refactor: extract runCompact from app.tsx to agent-turn.ts
Move the /compact command implementation into src/app/agent-turn.ts
with a typed RunCompactCtx interface. App.tsx keeps a thin useCallback
wrapper (~25 lines).

Commit-Queue-Session: cq_20260508160558_dad111b7260fef0bcd71f2b9
Coding-Agent: pi
Coding-Agent-Session: pi-019e06b6-830e-72ca-8378-e2ba2d3209c0
